+++ django_snowflake/base.py
@@ -91,7 +91,7 @@ class DatabaseWrapper(BaseDatabaseWrapper):
 
     def get_connection_params(self):
         settings_dict = self.settings_dict
-        conn_params = {}
+        conn_params = {'interpolate_empty_sequences':  True}
 
         if settings_dict['NAME']:
             conn_params['database'] = self.ops.quote_name(settings_dict['NAME'])